+++ NEWS.md
@@ -74,6 +74,7 @@
 * Cover contributor-types, contributor-name-types, instance-formats, instance-types, nature-of-content-terms, classification-types, modes-of-issuance APIs with integration tests ([MODINVSTOR-1166](https://folio-org.atlassian.net/browse/MODINVSTOR-1166))
 * Cover instance-statuses, electronic-access-relationships, statistical-code-types, statistical-codes, ill-policies, holdings-types, call-number-types APIs with integration tests ([MODINVSTOR-1167](https://folio-org.atlassian.net/browse/MODINVSTOR-1167))
 * Re-work all s3 related envs var and standardization ([MODINVSTOR-1403](https://folio-org.atlassian.net/browse/MODINVSTOR-1403))
+* Randomize sample data instance id/hrid and update references ([MODINVSTOR-1485](https://folio-org.atlassian.net/browse/MODINVSTOR-1485))
 
 ### Dependencies
 * Bump `LIB_NAME` from `OLD_VERSION` to `NEW_VERSION`